refactor: separate RabbitMQ and Kafka into dedicated NuGet packages
- Extract RabbitMQ implementation to MessageWorkerPool.RabbitMQ package
- Extract Kafka implementation to MessageWorkerPool.Kafka package
- Remove RabbitMQ/Kafka dependencies from core MessageWorkerPool package
- Move extension methods to respective package namespaces
- Update solution to include new project references
- Update CI/CD workflow to support multi-package tagging (rabbitmq-, kafka- prefixes)
- Update README with separate installation instructions for each package
- Fix namespace references across all test files
- Update example worker projects to use new package structure

BREAKING CHANGE: Namespace changes from MessageWorkerPool.RabbitMq to MessageWorkerPool.RabbitMQ
and MessageWorkerPool.KafkaMq to MessageWorkerPool.Kafka. Users must update using statements
and install the appropriate MessageWorkerPool.RabbitMQ or MessageWorkerPool.Kafka package.
