+++ src/MessageWorkerPool.sln
@@ -7,6 +7,10 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MessageWorkerPool", "Messag
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MessageWorkerPool.OpenTelemetry", "MessageWorkerPool.OpenTelemetry\MessageWorkerPool.OpenTelemetry.csproj", "{2A3B4C5D-6E7F-8901-2345-6789ABCDEF01}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MessageWorkerPool.RabbitMQ", "MessageWorkerPool.RabbitMQ\MessageWorkerPool.RabbitMQ.csproj", "{3B4C5D6E-7F89-0123-4567-89ABCDEF0123}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MessageWorkerPool.Kafka", "MessageWorkerPool.Kafka\MessageWorkerPool.Kafka.csproj", "{4C5D6E7F-8901-2345-6789-ABCDEF012345}"
+EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MessageWorkerPool.Test", "tests\MessageWorkerPool.Test\MessageWorkerPool.Test.csproj", "{35FA9645-24F8-4E20-94EB-0BED7ACA0D8B}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{CCB7A0F2-C13A-4838-87D1-76F275F8123C}"
@@ -28,6 +32,14 @@ Global
 		{2A3B4C5D-6E7F-8901-2345-6789ABCDEF01}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2A3B4C5D-6E7F-8901-2345-6789ABCDEF01}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2A3B4C5D-6E7F-8901-2345-6789ABCDEF01}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3B4C5D6E-7F89-0123-4567-89ABCDEF0123}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3B4C5D6E-7F89-0123-4567-89ABCDEF0123}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3B4C5D6E-7F89-0123-4567-89ABCDEF0123}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3B4C5D6E-7F89-0123-4567-89ABCDEF0123}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4C5D6E7F-8901-2345-6789-ABCDEF012345}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4C5D6E7F-8901-2345-6789-ABCDEF012345}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4C5D6E7F-8901-2345-6789-ABCDEF012345}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4C5D6E7F-8901-2345-6789-ABCDEF012345}.Release|Any CPU.Build.0 = Release|Any CPU
 		{35FA9645-24F8-4E20-94EB-0BED7ACA0D8B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{35FA9645-24F8-4E20-94EB-0BED7ACA0D8B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{35FA9645-24F8-4E20-94EB-0BED7ACA0D8B}.Release|Any CPU.ActiveCfg = Release|Any CPU
