+++ src/MessageWorkerPool/IWorker.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
-using RabbitMQ.Client.Events;
 
 namespace MessageWorkerPool
 {
