+++ src/MessageWorkerPool/Properties/AssemblyInfo.cs
@@ -1,0 +1,5 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("MessageWorkerPool.RabbitMQ")]
+[assembly: InternalsVisibleTo("MessageWorkerPool.Kafka")]
+[assembly: InternalsVisibleTo("MessageWorkerPool.Test")]
